Extract `FeedCachePolicy` to its own file
